Prepare campaign for C21_WMDE_Test_EN_04
Ticket: https://phabricator.wikimedia.org/T297703
