+++ campaign_info.toml
@@ -151,20 +151,20 @@ tracking = "wpde-mob01-211103-var"
 
 # English Campaign
 [english]
-campaign_tracking = "en03-ba-211126"
-preview_link = "/wiki/Wikipedia:Main_Page?devbanner={{banner}}&banner=B17WMDE_webpack_prototype"
+campaign_tracking = "en04-ba-211215"
+preview_link = "/wiki/Main_Page?devbanner={{banner}}&banner=B17WMDE_webpack_prototype"
 preview_url = 'https://en.wikipedia.org/wiki/Wikipedia:Main_Page?banner={{PLACEHOLDER}}'
 wrapper_template = "wikipedia_org"
 
 [english.banners.ctrl]
 filename = "./english/banner_ctrl.js"
-pagename = "B21_WMDE_Test_EN_03_ctrl"
-tracking = "org-en03-211126-ctrl"
+pagename = "B21_WMDE_Test_EN_04_ctrl"
+tracking = "org-en04-211215-ctrl"
 
 [english.banners.var]
 filename = "./english/banner_var.js"
-pagename = "B21_WMDE_Test_EN_03_var"
-tracking = "org-en03-211126-var"
+pagename = "B21_WMDE_Test_EN_04_var"
+tracking = "org-en04-211215-var"
 
 [english.test_matrix]
 platform = ["edge", "ie11", "firefox_win10", "chrome_win10", "safari", "firefox_macos", "chrome_macos", "firefox_linux", "chrome_linux"]
