+++ english/banner_var.js
@@ -1,17 +1,17 @@
 // eslint-disable-next-line no-unused-vars
-import style from './styles_var/styles.pcss';
+import style from './styles/styles.pcss';
 
 import * as formatters from '../shared/number_formatter/en';
 import { createCampaignParameters } from '../shared/campaign_parameters';
 import { createTrackingData } from '../shared/tracking_data';
 import { getTrackingIds } from '../shared/tracking_ids';
 
-import Banner from './components_var/Banner';
+import Banner from './components/Banner';
 import BannerPresenter from '../shared/banner_presenter';
 import Translations from '../shared/messages/en';
 import LocalTranslations from './translations';
-import DonationForm from './components_var/ui/form/DonationForm';
-import BannerText from './components_var/BannerText';
+import DonationForm from './components/ui/form/DonationForm';
+import BannerText from './components/BannerText';
 import Footer from '../shared/components/ui/EasySelectFooter';
 import useOfFundsText from '../node_modules/fundraising-frontend-content/i18n/en_GB/data/use_of_funds_content.json';
 import { createCampaignProjection } from '../shared/campaign_projection';
