+++ english/components/Banner.jsx
@@ -2,46 +2,28 @@ import { Component, h, createRef } from 'preact';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import BannerTransition from '../../shared/components/BannerTransition';
-import ProgressBar from '../../shared/components/ui/ProgressBar';
-import Footer from '../../shared/components/ui/EasySelectFooter';
-import Infobox from '../../shared/components/ui/Infobox';
-import FundsModal from '../../shared/components/ui/use_of_funds/FundsModal';
+import Infobox from './ui/Infobox';
 import FundsDistributionInfo from '../../shared/components/ui/use_of_funds/FundsDistributionInfo';
+import FundsModal from '../../shared/components/ui/use_of_funds/FundsModal';
 import TranslationContext from '../../shared/components/TranslationContext';
-import { CampaignProjection } from '../../shared/campaign_projection';
-import { LocalImpressionCount } from '../../shared/local_impression_count';
 import { BannerType } from '../../shared/BannerType';
+import createDynamicCampaignText from '../create_dynamic_campaign_text';
+import ProgressBar from '../../shared/components/ui/ProgressBar';
 
-const PENDING = 0;
-const VISIBLE = 1;
-const CLOSED = 2;
+const BannerVisibilityState = Object.freeze( {
+	PENDING: Symbol( 'pending' ),
+	VISIBLE: Symbol( 'visible' ),
+	CLOSED: Symbol( 'closed' )
+} );
 
 export default class Banner extends Component {
 
 	static propTypes = {
-		bannerName: PropTypes.string,
-		campaignName: PropTypes.string,
-
-		campaignParameters: PropTypes.object,
-		campaignProjection: PropTypes.instanceOf( CampaignProjection ),
-		formatters: PropTypes.object,
-		bannerText: PropTypes.elementType,
-		translations: PropTypes.object,
-		/**
-		 *  Contains items with values and labels for SelectGroup components
-		 *  in the form, e.g. amounts, payment types, intervals, etc
-		 */
-		formItems: PropTypes.object,
-
-		/** object that holds any data needed for tracking purposes */
-		trackingData: PropTypes.object,
-		impressionCounts: PropTypes.instanceOf( LocalImpressionCount ),
-
 		/** callback when banner closes */
 		onClose: PropTypes.func,
 		/** callback when banner gets submitted */
 		onSubmit: PropTypes.func,
-		/** Callback to register a displayBanner function with the BannerPresenter */
+		/** */
 		registerDisplayBanner: PropTypes.func.isRequired
 	}
 
@@ -50,19 +32,26 @@ export default class Banner extends Component {
 	constructor( props ) {
 		super( props );
 		this.state = {
-			displayState: PENDING,
+			bannerVisibilityState: BannerVisibilityState.PENDING,
 			isFundsModalVisible: false,
 
 			// trigger for banner resize events
 			formInteractionSwitcher: false
 		};
 		this.slideInBanner = () => {};
+		this.dynamicCampaignText = createDynamicCampaignText(
+			props.campaignParameters,
+			props.campaignProjection,
+			props.impressionCounts,
+			props.formatters,
+			props.translations
+		);
 	}
 
 	componentDidMount() {
 		this.props.registerDisplayBanner(
 			() => {
-				this.setState( { displayState: VISIBLE } );
+				this.setState( { bannerVisibilityState: BannerVisibilityState.VISIBLE } );
 				this.slideInBanner();
 			}
 		);
@@ -82,15 +71,15 @@ export default class Banner extends Component {
 	}
 
 	onFinishedTransitioning = () => {
-		this.startProgressbar();
 		this.props.onFinishedTransitioning();
+		this.startProgressbar();
 	}
 
 	closeBanner = e => {
 		e.preventDefault();
-		this.setState( { displayState: CLOSED } );
+		this.setState( { bannerVisibilityState: BannerVisibilityState.CLOSED } );
 		this.props.onClose();
-	}
+	};
 
 	registerBannerTransition = ( cb ) => {
 		this.slideInBanner = cb;
@@ -98,33 +87,35 @@ export default class Banner extends Component {
 
 	registerStartProgressbar = ( startPb ) => {
 		this.startProgressbar = startPb;
-	}
+	};
 
 	toggleFundsModal = () => {
 		if ( !this.state.isFundsModalVisible ) {
 			this.props.trackingData.tracker.trackBannerEvent( 'application-of-funds-shown', 0, 0, this.props.trackingData.bannerClickTrackRatio );
 		}
 		this.setState( { isFundsModalVisible: !this.state.isFundsModalVisible } );
-	}
+	};
 
 	fundsModalDonate = () => {
 		this.props.trackingData.tracker.trackBannerEvent( 'funds-modal-donate-clicked', 0, 0, this.props.trackingData.bannerClickTrackRatio );
 		this.setState( { isFundsModalVisible: false } );
-	}
+	};
 
 	onFormInteraction = () => {
-		this.setState( { formInteractionSwitcher: !this.state.formInteractionSwitcher } );
+		this.setState( { showLanguageWarning: true, formInteractionSwitcher: !this.state.formInteractionSwitcher } );
 	}
 
 	// eslint-disable-next-line no-unused-vars
 	render( props, state, context ) {
 		const DonationForm = props.donationForm;
 		const campaignProjection = props.campaignProjection;
+		const Footer = props.footer;
+
 		return <div
 			className={ classNames( {
 				'wmde-banner': true,
-				'wmde-banner--hidden': state.displayState === CLOSED,
-				'wmde-banner--visible': state.displayState === VISIBLE,
+				'wmde-banner--hidden': state.bannerVisibilityState === BannerVisibilityState.CLOSED,
+				'wmde-banner--visible': state.bannerVisibilityState === BannerVisibilityState.VISIBLE,
 				'wmde-banner--ctrl': props.bannerType === BannerType.CTRL,
 				'wmde-banner--var': props.bannerType === BannerType.VAR
 			} ) }
@@ -149,15 +140,11 @@ export default class Banner extends Component {
 										campaignParameters={props.campaignParameters}
 										campaignProjection={props.campaignProjection}
 										bannerText={props.bannerText}
-										propsForText={ { overallImpressionCount: props.impressionCounts.getOverallCount() } }
-									/>
-									<ProgressBar
-										formatters={props.formatters}
-										daysLeft={campaignProjection.getRemainingDays()}
-										donationAmount={campaignProjection.getProjectedDonationSum()}
-										goalDonationSum={campaignProjection.goalDonationSum}
-										missingAmount={campaignProjection.getProjectedRemainingDonationSum()}
-										setStartAnimation={ this.registerStartProgressbar }/>
+										dynamicCampaignText={ this.dynamicCampaignText }
+										propsForText={ {
+											overallImpressionCount: props.impressionCounts.getOverallCount(),
+											millionImpressionsPerDay: props.campaignParameters.millionImpressionsPerDay
+										} }/>
 								</div>
 							</div>
 							<div className="banner__form">
@@ -168,13 +155,23 @@ export default class Banner extends Component {
 									formatters={props.formatters}
 									impressionCounts={props.impressionCounts}
 									onFormInteraction={this.onFormInteraction}
-									customAmountPlaceholder={ props.translations[ 'custom-amount-placeholder' ] }
 									onSubmit={props.onSubmit}
-									locale="en_GB"
+									customAmountPlaceholder={ props.translations[ 'custom-amount-placeholder' ] }
+									buttonText={ props.buttonText }
+									errorPosition={ props.errorPosition }
+									bannerType={ props.bannerType }
+									showCookieBanner={ props.showCookieBanner }
 								/>
 							</div>
 						</div>
 						<Footer showFundsModal={ this.toggleFundsModal }/>
+						<ProgressBar
+							formatters={props.formatters}
+							daysLeft={campaignProjection.getRemainingDays()}
+							donationAmount={campaignProjection.getProjectedDonationSum()}
+							goalDonationSum={campaignProjection.goalDonationSum}
+							missingAmount={campaignProjection.getProjectedRemainingDonationSum()}
+							setStartAnimation={this.registerStartProgressbar}/>
 					</div>
 				</TranslationContext.Provider>
 			</BannerTransition>
@@ -183,11 +180,11 @@ export default class Banner extends Component {
 				onCallToAction={ this.fundsModalDonate }
 				isFundsModalVisible={ this.state.isFundsModalVisible }
 				useOfFundsText={ props.useOfFundsText }
-				locale='en'>
+				locale='de'>
 				<FundsDistributionInfo
-					applicationOfFundsData={ props.useOfFundsText.applicationOfFundsData } />
+					applicationOfFundsData={ props.useOfFundsText.applicationOfFundsData }
+				/>
 			</FundsModal>
 		</div>;
 	}
-
 }
