--- english/styles/BackLink.pcss
@@ -1,39 +1,0 @@
-$leftArrow: url( 'https://upload.wikimedia.org/wikipedia/commons/b/b7/Chevron-left.svg' );
-
-.wmde-banner {
-	.back {
-		position: absolute;
-		display: block;
-		left: 0;
-		top: 0;
-		height: 100%;
-		width: 20px;
-		margin-bottom: 10px;
-
-		@media ( min-width: $breakpoint_l ) {
-			position: relative;
-			height: 20px;
-		}
-
-		&__arrow {
-			position: absolute;
-			top: 50%;
-			left: 50%;
-			margin-top: -10px;
-			margin-left: -6px;
-			background: $leftArrow;
-			width: 12px;
-			height: 20px;
-			display: block;
-
-			&:hover {
-				opacity: 0.5;
-			}
-
-			@media ( min-width: $breakpoint_l ) {
-				left: 0;
-				margin-left: 0;
-			}
-		}
-	}
-}
