+++ english/styles/ProgressBar.pcss
@@ -1,11 +1,20 @@
+$progress-bar-height: 30px;
+
 .wmde-banner {
 	.progress_bar {
 		display: flex;
 		font-family: $banner-main-font-family;
-		font-size: 14px;
+		font-size: 12px;
+		line-height: $progress-bar-height;
+
+		&__days_left,
+		&__donation_text,
+		&__donation_remaining {
+			opacity: 0;
+			visibility: hidden;
+		}
 
 		&__days_left {
-			visibility: hidden;
 			font-weight: bold;
 			padding-left: 5px;
 
@@ -15,25 +24,25 @@
 		}
 		&__wrapper {
 			position: relative;
+			height: $progress-bar-height;
 			display: flex;
 			justify-content: space-between;
 			flex: 1;
-			border: 2px solid $color-primary-light;
-			background-color: $color-white;
-			border-radius: 0.55em;
-			margin-top: 1em;
+			border-top: 1px solid $color-secondary-light;
+			border-bottom: 1px solid $color-secondary-light;
+			background-color: $color-secondary-lighter;
 		}
 		&__donation_fill {
 			width: 0;
 			min-width: 5em;
+			max-width: 100%;
 			display: flex;
 			justify-content: space-between;
 
 			background: $color-secondary;
-			height: 22px;
-			border-radius: 0.4em;
-			transition-duration: 3000ms;
-			transition-property: width;
+			height: $progress-bar-height;
+			border-radius: 0 15px 15px 0;
+			transition: width 3000ms $banner-easing;
 		}
 		&__donation_text {
 			text-align: right;
@@ -41,11 +50,10 @@
 			overflow: hidden;
 			padding-right: 16px;
 			flex: 1;
-			visibility: hidden;
-
 			color: $color-white;
-			line-height: 1.8em;
 			white-space: nowrap;
+
+			transition: opacity 300ms $banner-easing;
 		}
 		&__donation_remaining {
 			padding: 0 10px;
@@ -54,8 +62,6 @@
 			white-space: nowrap;
 
 			color: $color-black;
-			line-height: 1.6em;
-			margin-left: -40px;
 		}
 		&__pointer_tip {
 			width: 5px;
@@ -78,16 +84,14 @@
 			position: relative;
 			font-weight: bold;
 
-			padding-top: 16px;
-			margin-left: -20px;
-			padding-right: 37px;
-			color: $color-white;
+			padding-top: 9px;
 		}
 
 		&--finished {
 			.progress_bar__days_left,
 			.progress_bar__donation_text,
 			.progress_bar__donation_remaining {
+				opacity: 1;
 				visibility: visible;
 			}
 		}
@@ -96,20 +100,6 @@
 			.progress_bar__days_left,
 			.progress_bar--finished .progress_bar__days_left {
 				display: none;
-			}
-		}
-
-		&--lateprogress {
-			.progress_bar__donation_remaining--inner {
-				display: none;
-			}
-
-			.progress_bar__donation_remaining--outer {
-				display: block;
-			}
-
-			.progress_bar__donation_fill {
-				margin-right: 16px;
 			}
 		}
 	}
