+++ english/styles/Slider.pcss
