+++ english/styles/styles.pcss
@@ -1,16 +1,16 @@
 @import '../../shared/components/ui/use_of_funds/variables.pcss';
-@import './variables_ctrl.pcss';
+@import 'variables.pcss';
 
 @import '../../shared/components/BannerTransition.pcss';
 
 @import './Banner.pcss';
-@import './DonationForm.pcss';
-@import './SelectGroup_ctrl.pcss';
+@import 'DonationForm.pcss';
+@import './SelectGroup.pcss';
 @import './Footer.pcss';
 @import './Infobox.pcss';
-@import './BannerText.pcss';
+@import 'BannerText.pcss';
 @import './ProgressBar.pcss';
-
+@import 'Slider.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsContent.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsDistributionInfo.pcss';
