--- english/styles/variables_ctrl.pcss
@@ -1,32 +1,0 @@
-$color-primary: #0166a5;
-$color-primary-light: #dce4e8;
-$color-secondary: #849d0e;
-$color-secondary-light: #dbe5e7;
-$color-secondary-hover: #97b313;
-$color-error:  #990000;
-
-$color-white: #ffffff;
-$color-black: #000000;
-$color-gray: #808080;
-$color-gray-medium: #747474;
-$color-black-light: #222222;
-$color-gray-light: #dddddd;
-
-$banner-main-font-family: 'Arial', 'Helvetica', 'Verdana', sans-serif;
-$banner-text-font-family: 'Source Serif', 'Linux Libertine', 'Linux Libertine O', 'Georgia', 'Bitstream Charter', serif;
-
-$form-width: 300px;
-$form-element-border-radius: 2px;
-
-$errorIcon: url( "data:image/svg+xml,%3Csvg width='16' height='16' viewBox='0 0 16 16' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M11.3138 0H4.68622L0 4.68622V11.3138L4.68622 16H11.3138L16 11.3138V4.68622L11.3138 0ZM8.88889 12.4444H7.11111V10.6667H8.88889V12.4444ZM8.88889 8.88889H7.11111V3.55556H8.88889V8.88889Z' fill='%23812923'/%3E%3C/svg%3E%0A" );
-
-/*
-Always work 'mobile first':
-	do not use @media ( max-width: $breakpoint-x )
-	always use @media ( min-width: $breakpoint-x )
-*/
-$breakpoint_xs: 900px;
-$breakpoint_s: 1024px;
-$breakpoint_m: 1090px;
-$breakpoint_l: 1300px;
-$breakpoint_xl: 1580px;
